Make all sizes of type `i64`.
This has wide-ranging implications for the types of things in the prelude:

* Functions like `replicate` and `iota` now take `i64` arguments.

* The `from_fraction` function now takes `i64`.

* The `to_i32` function has been removed.

Closes #134.

+++ rts/python/opencl.py
@@ -122,7 +122,7 @@ def initialise_opencl_object(self,
 
     self.global_failure = self.pool.allocate(np.int32().itemsize)
     cl.enqueue_fill_buffer(self.queue, self.global_failure, np.int32(-1), 0, np.int32().itemsize)
-    self.global_failure_args = self.pool.allocate(np.int32().itemsize *
+    self.global_failure_args = self.pool.allocate(np.int64().itemsize *
                                                   (self.global_failure_args_max+1))
     self.failure_is_an_option = np.int32(0)
 
@@ -225,7 +225,7 @@ def sync(self):
         cl.enqueue_fill_buffer(self.queue, self.global_failure, np.int32(-1), 0, np.int32().itemsize)
 
         # Read failure args.
-        failure_args = np.empty(self.global_failure_args_max+1, dtype=np.int32)
+        failure_args = np.empty(self.global_failure_args_max+1, dtype=np.int64)
         cl.enqueue_copy(self.queue, failure_args, self.global_failure_args, is_blocking=True)
 
         raise Exception(self.failure_msgs[failure[0]].format(*failure_args))
