+++ src/Language/Futhark/Prop.hs
@@ -832,8 +832,8 @@ intrinsics =
              ( "unflatten",
                IntrinsicPolyFun
                  [tp_a]
-                 [ Scalar $ Prim $ Signed Int32,
-                   Scalar $ Prim $ Signed Int32,
+                 [ Scalar $ Prim $ Signed Int64,
+                   Scalar $ Prim $ Signed Int64,
                    Array () Nonunique t_a (rank 1)
                  ]
                  $ Array () Nonunique t_a (rank 2)
@@ -847,7 +847,7 @@ intrinsics =
              ( "rotate",
                IntrinsicPolyFun
                  [tp_a]
-                 [Scalar $ Prim $ Signed Int32, arr_a]
+                 [Scalar $ Prim $ Signed Int64, arr_a]
                  arr_a
              ),
              ("transpose", IntrinsicPolyFun [tp_a] [arr_2d_a] arr_2d_a),
@@ -855,7 +855,7 @@ intrinsics =
                IntrinsicPolyFun
                  [tp_a]
                  [ Array () Unique t_a (rank 1),
-                   Array () Nonunique (Prim $ Signed Int32) (rank 1),
+                   Array () Nonunique (Prim $ Signed Int64) (rank 1),
                    Array () Nonunique t_a (rank 1)
                  ]
                  $ Array () Unique t_a (rank 1)
@@ -865,11 +865,11 @@ intrinsics =
              ( "hist",
                IntrinsicPolyFun
                  [tp_a]
-                 [ Scalar $ Prim $ Signed Int32,
+                 [ Scalar $ Prim $ Signed Int64,
                    uarr_a,
                    Scalar t_a `arr` (Scalar t_a `arr` Scalar t_a),
                    Scalar t_a,
-                   Array () Nonunique (Prim $ Signed Int32) (rank 1),
+                   Array () Nonunique (Prim $ Signed Int64) (rank 1),
                    arr_a
                  ]
                  uarr_a
@@ -897,28 +897,28 @@ intrinsics =
                IntrinsicPolyFun
                  [tp_a]
                  [ Scalar (Prim $ Signed Int32),
-                   Scalar t_a `arr` Scalar (Prim $ Signed Int32),
+                   Scalar t_a `arr` Scalar (Prim $ Signed Int64),
                    arr_a
                  ]
-                 $ tupleRecord [uarr_a, Array () Unique (Prim $ Signed Int32) (rank 1)]
+                 $ tupleRecord [uarr_a, Array () Unique (Prim $ Signed Int64) (rank 1)]
              ),
              ( "map_stream",
                IntrinsicPolyFun
                  [tp_a, tp_b]
-                 [Scalar (Prim $ Signed Int32) `karr` (arr_ka `arr` arr_kb), arr_a]
+                 [Scalar (Prim $ Signed Int64) `karr` (arr_ka `arr` arr_kb), arr_a]
                  uarr_b
              ),
              ( "map_stream_per",
                IntrinsicPolyFun
                  [tp_a, tp_b]
-                 [Scalar (Prim $ Signed Int32) `karr` (arr_ka `arr` arr_kb), arr_a]
+                 [Scalar (Prim $ Signed Int64) `karr` (arr_ka `arr` arr_kb), arr_a]
                  uarr_b
              ),
              ( "reduce_stream",
                IntrinsicPolyFun
                  [tp_a, tp_b]
                  [ Scalar t_b `arr` (Scalar t_b `arr` Scalar t_b),
-                   Scalar (Prim $ Signed Int32) `karr` (arr_ka `arr` Scalar t_b),
+                   Scalar (Prim $ Signed Int64) `karr` (arr_ka `arr` Scalar t_b),
                    arr_a
                  ]
                  $ Scalar t_b
@@ -927,7 +927,7 @@ intrinsics =
                IntrinsicPolyFun
                  [tp_a, tp_b]
                  [ Scalar t_b `arr` (Scalar t_b `arr` Scalar t_b),
-                   Scalar (Prim $ Signed Int32) `karr` (arr_ka `arr` Scalar t_b),
+                   Scalar (Prim $ Signed Int64) `karr` (arr_ka `arr` Scalar t_b),
                    arr_a
                  ]
                  $ Scalar t_b
